==> 3_oil_price_src _clean.ipynb <==
import pandas as pd
import datetime
pd.set_option('precision', 3)
pd.set_option('max_rows', 20)
pd.set_option('max_colwidth', 30)

%config Completer.use_jedi = False

## 1.  cleaning 'oil_price_daily_src.csv'


#read the daily file
daily = pd.read_csv("oil_price_daily_src.csv", 
                                   encoding='ISO-8859-1')

#show the dataset
daily.head(15)

#show null value =>no null value
daily[daily.isnull().values==True]

### correct the column 'Date'

# remane the column name
daily.columns = ['Date','Open','High','Low','Close']

# transfer to data frame
pd.DataFrame(daily)

# change the type of 'Date'
daily['Date'].astype(str)


# correct the column 'Date'
daily['Date'][:]=daily['Date'][:].str[-10:]


# set the column 'Date' as time format
daily['Date'] = pd.to_datetime(daily['Date'],format='%m/%d/%Y').dt.normalize()


# sort the data by 'Date'
daily.sort_values(by=['Date'],inplace = True) 
daily.tail(20)

### strip '$'

# strip '$'
daily['Open'] = daily['Open'].str.replace('$', '').astype(float)
daily['High'] = daily['High'].str.replace('$', '').astype(float)
daily['Low'] = daily['Low'].str.replace('$', '').astype(float)
daily['Close'] = daily['Close'].str.replace('$', '').astype(float)


# export the csv file
daily.to_csv('oil_price_daily_stage.csv', index=False)
daily

## 2. cleaning 'oil_price_monthly_src.csv'

### follow the same process as above

#read the monthly file
monthly = pd.read_csv("oil_price_monthly_src.csv", 
                                   encoding='ISO-8859-1')

#show null value =>no null value
monthly[monthly.isnull().values==True]

# remane the column name
monthly.columns = ['Date','Open','High','Low','Close']

# transfer to data frame
pd.DataFrame(monthly)

# change the type of 'Date'
monthly['Date'].astype(str)


# correct the column 'Date'
monthly['Date'][:]=monthly['Date'][:].str[-7:]

# strip '$'
monthly['Open'] = monthly['Open'].str.replace('$', '').astype(float)
monthly['High'] = monthly['High'].str.replace('$', '').astype(float)
monthly['Low'] = monthly['Low'].str.replace('$', '').astype(float)
monthly['Close'] = monthly['Close'].str.replace('$', '').astype(float)

# export the csv file
monthly.to_csv('oil_price_monthly_stage.csv', index=False)
