==> 4_carsales_src _dirty_clean.ipynb <==
import pandas as pd

pd.set_option('precision', 3)
pd.set_option('max_rows', 20)
pd.set_option('max_colwidth', 30)

%config Completer.use_jedi = False

## 1.  cleaning general-motors_carsales_src _dirty.csv


#read the GM file
gm = pd.read_csv("general-motors_carsales_src _dirty.csv", 
                                   encoding='ISO-8859-1')

#show null value
gm[gm.isnull().values==True]

### fill the null value in 'Year'

# check the rows near the null value (index 41)
print(gm.iloc[35:45])

# from the result, we know the null value should be 2008
# fill in '2008'
gm['Year']=gm['Year'].fillna(2008)

# change the datatype of 'year' from float to intger 
gm['Year']=gm['Year'].astype(int)

### fill the null value in 'Month'

# check the rows near the null value (index 26)
print(gm.iloc[20:30])

# from the result, we know the null value is 'Mar', so we fill 'Mar' into the cell
gm['Month'][26]='Mar'

### fix the null value in index 62

# check the row near index 62
print(gm.iloc[60:65])

# the data in column 'Month' is combined with column 'general-motors' 
# the string is split into two parts
gm['Month'][62]=gm['general-motors'][62][0:3]
gm['general-motors'][62]=gm['general-motors'][62][3:]

# change the datatype of 'year' from object to intger 
gm['general-motors']=gm['general-motors'].astype(float)

### fix the null value in 'general-motors'

# list all value & mean of 2005
print(gm.loc[gm['Year']==2005])
print('*'*20)
mean_2005=gm.loc[gm['Year']==2005].mean()
print('The mean value of 2005:',mean_2005)
print('*'*20)

# the all value and mean of Feb from 2005-2015
print(gm.loc[gm['Month']=='Feb'])
print('*'*20)
mean_2005_2015_feb=gm['general-motors'].loc[gm['Year'] <= 2015].groupby([gm['Month']]).mean()['Feb']
print('The mean value of Feb from 2005-2010: ',mean_2005_2015_feb)

# fill the null value using mean of Feb 2005-2015
gm['general-motors'][1]=mean_2005_2015_feb

#check null value again
print('Current number of null value in the dataset:', gm.isnull().any().sum())

### fix the outlier value

# check total data
pd.set_option('display.float_format',lambda x : '%.3f' % x)
gm['general-motors'].describe()

# an extra large number is observed
max=gm['general-motors'].max()
print(gm.loc[gm['general-motors'] == max])

# '1' in string'1395217.00' is redundant, change to '395217'
gm['general-motors'][17]=395217

# another outlier is observed. The number should not have decimal places.
print(gm['general-motors'].nsmallest(15,keep='all'))

# it is believed that '189.561' is error, it should be '189561'
gm['general-motors'][52]=189561

# check is any duplicated row =>0
gm[gm.duplicated()]

# remane the column name
gm=gm.rename(columns={'general-motors':'Gm'})

# change the datatype of 'Gm' to intger 
gm['Gm']=gm['Gm'].astype(int)

## 2.  cleaning ford-motor-company_carsales_src _dirty.csv

#read the Ford file
ford = pd.read_csv("ford-motor-company_carsales_src _dirty.csv", 
                                   encoding='ISO-8859-1')

#show null value
ford[ford.isnull().values==True]

### drop the row that all cells are null value

#drop the row that all values are null values
ford=ford.dropna(axis=0, how='all')

# check the left rows which contain null value
ford[ford.isnull().values==True]

### fix the  rows with error

# fix the rows with error
ford['ford-motor-company'][3]=ford['Year'][3][7:]
ford['Month'][3]=ford['Year'][3][4:7]
ford['Year'][3]=ford['Year'][3][0:4]

#check if the error fix
ford.iloc[3:5]

### drop the duplicated row

# check the duplicated rows
ford[ford.duplicated()]

# drop the duplicated row
ford.drop_duplicates(inplace = True)

# check if duplicated row still exist
ford[ford.duplicated()]

### sort the rows by month

# original row : ordered, changed row: Jan and Mar are exchanged
ford[24:36]

# create a new column and sort by 'Year' and 'Month', then drop 'month_number'
ford['month_number']=ford['Month'].replace(['Jan','Feb','Mar','Apr','May','Jun','Jul','Aug','Sep','Oct','Nov','Dec'],[1,2,3,4,5,6,7,8,9,10,11,12])
ford=ford.sort_values(by=['Year','month_number'])
ford=ford.drop(['month_number'],axis=1)
ford[24:36]

# remane the column name
ford=ford.rename(columns={'ford-motor-company':'Ford'})

# change the datatype of 'Ford' to intger 
ford['Ford']=ford['Ford'].astype(int)

## 3. cleaning tesla-inc_carsales_src _dirty.csv

#read the Ford file
tesla = pd.read_csv("tesla-inc_carsales_src _dirty.csv", 
                                   encoding='ISO-8859-1')

#show null value
tesla[tesla.isnull().values==True]

### fix the moved row

# it is observed that row (index 2) is shifted
tesla

# fix the moved row
tesla['Year'][1]=tesla['Month'][1]
tesla['Month'][1]=tesla['tesla-inc'][1]
tesla['tesla-inc'][1]=tesla['Unnamed: 3'][1]
tesla=tesla.drop(['Unnamed: 3'],axis=1)
tesla.head(10)

### fix the abnormal string

# show the abnormal string in 'Year', '20 18' is found
print(tesla['Year'].value_counts())
print('*'*20)
print(tesla.loc[tesla['Year']=='20 18'])

# replace the abnormal string
tesla['Year'][43]=tesla['Year'][43].replace(' ','')

# show the abnormal string in 'Month', 'Jan%' is found
print(tesla['Month'].value_counts())
print('*'*20)
print(tesla.loc[tesla['Month']=='Jan%'])

# replace the abnormal string
tesla['Month'][24]=tesla['Month'][24].replace('%','')

# remane the column name
tesla=tesla.rename(columns={'tesla-inc':'Tesla'})

# change the datatype of 'Tesla' to intger 
tesla['Tesla']=tesla['Tesla'].astype(int)

## merge three dataset

# create a new column which combines 'Year' and 'Month', and remove 'Year' and 'Month'
def process(name):
    name['Year']=name['Year'].astype(str)
    name['year_month'] = name['Year'] + "." + name['Month']
    name = name.drop(labels=['Year','Month'],axis=1)
    cols_name = name.columns.tolist()
    cols_name = cols_name[-1:] + cols_name[:-1]
    cols_name
    name = name[cols_name]
    return name

GM1=process(gm)
FORD1=process(ford)
TESLA1=process(tesla)

# merge the dataset
total = pd.merge(GM1,FORD1,how='left',on = 'year_month')
total1 = pd.merge(total,TESLA1,how='left',on = 'year_month')

# instead null value with 0 in total1['Tesla']
total1['Tesla']=total1['Tesla'].fillna(0)
total1['Tesla']=total1['Tesla'].astype(int)

# outpot file
total1.to_csv("carsales_stage.csv", index=False)

total1
